[idemp-machine] Add a checkpoint reason for marker
Optionally we can add marker function calls that can be interpreted by the
emulator to gather statistics about why a checkpoint was inserted, how many
times it was executed etc.

Running code with these markers not patched out (i.e, on real hardware or
without using the emulator plugin that skips them) will lead to crashes
as the LR register is not saved before the call.