[llvm][epilog-opt-pass] Add pass
This pass disables the interrupts when multiple POP instructions (or
other STACK shrink operations) occur in the epilog.
Normally each of them would require a checkpoint so that a potential
ISR will not cause a WAR. This pass disables the interrupts before the
sequence of SP modifications and enables them again after.
This is combined with a special checkpoint code implementation that does
not use any stack memory (next commit)